requirements.txt
mysql-connector-python==8.0.28
boto3==1.22.4
flask==2.1.1
pyjwt==2.3.0

